--HG--
branch : yt

+++ yt/data_objects/hierarchy.py
@@ -354,7 +354,7 @@ class AMRHierarchy(ObjectFindingMixin, ParallelAnalysisInterface):
         #   1 = number of cells
         #   2 = blank
         desc = {'names': ['numgrids','numcells','level'],
-                'formats':['Int32']*3}
+                'formats':['Int64']*3}
         self.level_stats = blankRecordArray(desc, MAXLEVEL)
         self.level_stats['level'] = [i for i in range(MAXLEVEL)]
         self.level_stats['numgrids'] = [0 for i in range(MAXLEVEL)]
